feat: version 1.3.0

+++ pack_brimstone.py
@@ -9,15 +9,32 @@ def main(args):
 	if not version:
 		return
 	target_dir = f"./dist/brimstone_{version}"
-	os.mkdir(target_dir)
-	for file in os.listdir("."):
-		if file.endswith(".ghuser"):
-			shutil.move(f"./{file}", f"{target_dir}/{file}")
-		elif file == f"brimstone_{version}.gh":
-			shutil.copy(f"./{file}", f"{target_dir}/{file}")
-	for file in os.listdir("./dist_template"):
-		shutil.copy(f"./dist_template/{file}", f"{target_dir}/{file}")
-	shutil.make_archive(target_dir, "zip", target_dir)
+	try:
+		os.mkdir(target_dir)
+		for file in os.listdir("."):
+			if file.endswith(".ghuser"):
+				shutil.move(f"./{file}", f"{target_dir}/{file}")
+			elif file == f"brimstone_{version}.gh":
+				shutil.copy(f"./{file}", f"{target_dir}/{file}")
+
+		for item in os.listdir("./dist_template"):
+			print(item)
+			if os.path.isfile(item):
+				file = item
+				shutil.copy(f"./dist_template/{file}", f"{target_dir}/{file}")
+			else:
+				folder = item
+				os.mkdir(f"{target_dir}/{folder}")
+				for file in os.listdir(f"./dist_template/{folder}"):
+					shutil.copy(f"./dist_template/{folder}/{file}", f"{target_dir}/{folder}/{file}")
+		shutil.make_archive(target_dir, "zip", target_dir)
+	except:
+		# it failed, remove the broken directory
+		shutil.rmtree(target_dir)
+		zip_file = f"{target_dir}.zip"
+		if os.path.isfile(zip_file):
+			os.remove(zip_file)
+		raise
 
 
 if __name__ == '__main__':
